Vard order controller

+++ ACA.DeliverySystem.Business/Services/ItemService.cs
@@ -16,37 +16,50 @@ namespace ACA.DeliverySystem.Business.Services
             _mapper = mapper;
         }
 
-        public async Task<Item> CreateItem(Item item, CancellationToken token)
+        public async Task<Item> CreateItem(ItemAddModel item, CancellationToken token)
         {
-            await _uow.ItemRepository.Add(item, token);
+            var mappedModel = _mapper.Map<Item>(item);
+            await _uow.ItemRepository.Add(mappedModel, token);
             await _uow.Save(token);
-            return item;
+            return mappedModel;
         }
 
         public async Task Delete(int id, CancellationToken token)
         {
+            var item = await _uow.ItemRepository.GetById(id, token);
+
+            if (item == null)
+                throw new KeyNotFoundException($"Item with ID {id} not found.");
+
+
             await _uow.ItemRepository.Delete(id, token);
             await _uow.Save(token);
         }
 
-        public async Task<IEnumerable<Item>> GetAll(CancellationToken token)
+        public async Task<IEnumerable<ItemViewModel>> GetAll(CancellationToken token)
         {
-            return await _uow.ItemRepository.GetAll(token);
+            var items = await _uow.ItemRepository.GetAll(token);
+            return _mapper.Map<IEnumerable<ItemViewModel>>(items);
         }
 
-        public async Task<Item> GetById(int id, CancellationToken token)
+        public async Task<ItemViewModel> GetById(int id, CancellationToken token)
         {
-            return await _uow.ItemRepository.GetById(id, token);
+            var item = await _uow.ItemRepository.GetById(id, token);
+            return _mapper.Map<ItemViewModel>(item);
         }
 
         public async Task Update(int id, ItemUpdateModel model, CancellationToken token)
         {
-            var oldItem = await GetById(id, token);
+            var oldItem = await _uow.ItemRepository.GetById(id, token);
+
+            if (oldItem == null)
+                throw new KeyNotFoundException($"Item with ID {id} not found.");
+
             // Code for review
             oldItem.Description = model.Description;
             oldItem.Price = model.Price;
             oldItem.Name = model.Name;
-            //oldItem = _mapper.Map<Item>(model);
+
             await _uow.ItemRepository.Update(oldItem, token);
             await _uow.Save(token);
         }
